Override _strip_protocol method (#297)
This change overrides the internal `_strip_protocol` method for the fs.
The current implementation uses the `AbstractFileSystem` method which
doesn't return the correct result for GCS urls. Instead, we override
it with `infer_storage_options` method that other remote filesystem
implementation uses.

Fixing this also fixes some other bugs that rely on this method. One of
such bugs was encountered when uploading a file to a folder in GCS and
the implementation before the change infered the remote folder as the
filename. I added a test to check for this bug and it works with the
fix.