Update cleanup logic to empty the bucket instead of deleting the bucket
When tests are running in real GCS, we will make sure the bucket is
empty before each test run. In case of emulator we will continue to
use the existing logic to delete and create the buckets.

In case of real gcs, we might end up in a race condition and tests fail
with FileNotFound error, this can happen if bucket creation gets
skipped after the deletion from previous test is delayed.

Creating and deleting bucket for each test hugely increases the runtime
of the test suite.